requirements.txt
alembic==1.11.1
anyio==3.7.1
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
click==8.1.5
dnspython==2.4.0
email-validator==1.0.5
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.13.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
psycopg2-binary==2.9.6
pytest==7.4.0
sniffio==1.3.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.3.6
WTForms==3.0.1
